Improve Hypothesis tests efficiency
With the previous way of creating lists of actions most of examples
seemed different to Hypothesis (because they have different integers),
but not to the build_from (because most of these integers were just
ignored). The new generator model requires additional parameter, but it
should generate more diverse lists of examples (because no generated
data is ignored now).