Unpin astroid dependency
A long time ago we had to pin astroid version because of bug [1] in the
current version. This bug led to false errors during 'make lint' check.
Nowadays, it's been perished. So, we no longer need to use any kind of
fixed astroid version. Just use one that comes with pylint itself.

[1]: https://github.com/PyCQA/pylint/issues/1609