Fix: improve rendering for the case of many children of a single node
For the given case:

A   B   C   D   E   F
  \  \  |  /  /   /
    \ \ |/  /  /
        G

All edges (G-A, G-B,..) was counted as a single "edge to G".  This
behavior inhibited a rule that is used to limit total amount of goals on
a layer.

Now we count each edge separately, so such situation is now impossible.
Children goal A-F will now be placed on 2 layers instead of 1.